+++ standalone-parquet/src/main/java/io/delta/standalone/util/ParquetSchemaConverter.java
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package io.delta.utils;
+package io.delta.standalone.util;
 
-import io.delta.utils.internal.SparkToParquetSchemaConverter;
 import org.apache.parquet.schema.MessageType;
 
 import io.delta.standalone.types.StructType;
+
+import io.delta.standalone.internal.util.SparkToParquetSchemaConverter;
 
 /**
  * :: DeveloperApi ::
